[cmd-pathlookup] Allow unqualified command names via path lookup
Prior to this patch the test-command argument was required to be an
absolute or relative path.
When I'm simply specifying a solver as test command, I usually don't
care where the binary is. This patch allows to use the basename
of the test command, and simply performs a path lookup to find
the binary. The behavior for absolute/relative paths should be
unchanged.

This allows to use commands like

	ddsmt in out sh -c 'solver $1 | grep interesting' --

Instead of having to use "/bin/sh ...".